elka-studio/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pyyaml
